requirements.txt
requests
python-dotenv
volcengine
pypinyin
playwright
openai
fal-client
